pyproject.toml
[project]
name = "week-2-totoo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.35.0",
]
